refactor: major cleanup, break out into sepereate folder/files
