+++ package.json
@@ -45,7 +45,6 @@
         "react-dom": "^18.2.0",
         "react-error-boundary": "^3.1.4",
         "react-hook-form": "^7.42.1",
-        "react-router-dom": "^6.10.0",
         "react-select": "^5.7.0",
         "react-shepherd": "^4.2.0",
         "swr": "^2.0.1"
