+++ src/NextGen/DonationForm/resources/app/form/MultiStepForm/components/SubmitButton.tsx
@@ -1,0 +1,13 @@
+import {__} from '@wordpress/i18n';
+
+export default function SubmitButton({
+    isSubmitting,
+    submittingText = __('Submitting…', 'give'),
+    buttonText = __('Donate Now', 'give'),
+}) {
+    return (
+        <button type="submit" disabled={isSubmitting} aria-busy={isSubmitting}>
+            {isSubmitting ? submittingText : buttonText}
+        </button>
+    );
+};
